Create a single test client
This simplifies unit tests and will make them easier to extend in future.
