+++ slack/examples_test.go
@@ -1,7 +1,6 @@
 package slack
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"strings"
@@ -9,25 +8,14 @@ import (
 
 	"github.com/slack-go/slack"
 	"github.com/slack-go/slack/slackevents"
-	"github.com/slack-go/slack/socketmode"
 	"github.com/theothertomelliott/spanner"
 )
 
 // TestGettingStarted verifies that the code in examples/gettingstarted
 // interacts with Slack in the expected way
 func TestGettingStarted(t *testing.T) {
-	slackEvents := make(chan socketmode.Event)
-	client := newExampleTestClient()
-
-	testApp := newAppWithClient(
-		client,
-		slackEvents,
-	)
-
-	postEvent := make(chan interface{}, 10)
-	testApp.SetPostEventFunc(func() {
-		postEvent <- struct{}{}
-	})
+	client := newTestClient()
+	testApp := client.CreateApp()
 
 	go func() {
 		err := testApp.Run(handler)
@@ -37,16 +25,13 @@ func TestGettingStarted(t *testing.T) {
 	}()
 
 	// Send hello message
-	slackEvents <- messageEvent(
+	client.SendEventToApp(messageEvent(
 		slackevents.MessageEvent{
 			Text:    "hello",
 			Channel: "ABC123",
 			User:    "DEF456",
 		},
-	)
-
-	// Wait for event to be handled
-	<-postEvent
+	))
 
 	// Expect a single message and clear the message list
 	if len(client.messagesSent) != 1 {
@@ -61,7 +46,7 @@ func TestGettingStarted(t *testing.T) {
 	}
 
 	// Select an option value
-	slackEvents <- messageInteractionEvent(
+	client.SendEventToApp(messageInteractionEvent(
 		"hash",
 		"timestamp",
 		msg.metadata,
@@ -77,10 +62,7 @@ func TestGettingStarted(t *testing.T) {
 				},
 			},
 		},
-	)
-
-	// Wait for event to be handled
-	<-postEvent
+	))
 
 	if len(client.messagesSent) != 1 {
 		t.Errorf("expected one message to be sent, got %d", len(client.messagesSent))
@@ -106,58 +88,4 @@ func handler(ev spanner.Event) error {
 		}
 	}
 	return nil
-}
-
-func newExampleTestClient() *exampleTestClient {
-	return &exampleTestClient{
-		stop: make(chan struct{}),
-	}
-}
-
-type exampleTestClient struct {
-	nilSocketClient
-
-	messagesSent    []sentMessage
-	messagesUpdated []updatedMessage
-
-	stop chan struct{}
-}
-
-type sentMessage struct {
-	channelID string
-	blocks    []slack.Block
-	metadata  slack.SlackMetadata
-}
-
-type updatedMessage struct {
-	sentMessage
-	timestamp string
-}
-
-func (r *exampleTestClient) RunContext(context.Context) error {
-	<-r.stop // Must block
-	return nil
-}
-
-func (*exampleTestClient) Ack(req socketmode.Request, payload ...interface{}) {}
-
-func (c *exampleTestClient) SendMessageWithMetadata(ctx context.Context, channelID string, blocks []slack.Block, metadata slack.SlackMetadata) (string, string, string, error) {
-	c.messagesSent = append(c.messagesSent, sentMessage{
-		channelID: channelID,
-		blocks:    blocks,
-		metadata:  metadata,
-	})
-	return "", "", "", nil
-}
-
-func (c *exampleTestClient) UpdateMessageWithMetadata(ctx context.Context, channelID string, timestamp string, blocks []slack.Block, metadata slack.SlackMetadata) (string, string, string, error) {
-	c.messagesUpdated = append(c.messagesUpdated, updatedMessage{
-		sentMessage: sentMessage{
-			channelID: channelID,
-			blocks:    blocks,
-			metadata:  metadata,
-		},
-		timestamp: timestamp,
-	})
-	return "", timestamp, "", nil
 }
