+++ slack/receivemessage_test.go
@@ -4,25 +4,19 @@ import (
 	"testing"
 
 	"github.com/slack-go/slack/slackevents"
-	"github.com/slack-go/slack/socketmode"
 	"github.com/theothertomelliott/spanner"
 )
 
 func TestReceiveMessageContent(t *testing.T) {
-	slackEvents := make(chan socketmode.Event, 10)
-	client := newRunSocketClient()
-
-	testApp := newAppWithClient(
-		client,
-		slackEvents,
-	)
+	client := newTestClient()
+	testApp := client.CreateApp()
 
 	message := slackevents.MessageEvent{
 		Channel: "ABC123",
 		User:    "DEF456",
 		Text:    "hello, there",
 	}
-	slackEvents <- messageEvent(message)
+	client.SendEventToAppAsync(messageEvent(message))
 
 	testApp.Run(func(evt spanner.Event) error {
 		defer func() {
